requirements.txt
apache-airflow==2.11.0

GitPython==3.1.44
nbconvert==7.16.6
nbformat==5.10.4

google-api-python-client==2.169.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
psutil==7.0.0
pyperclip==1.9.0
python-dotenv==1.1.0
selenium==4.28.1

pillow==11.1.0
CairoSVG==2.7.1

polars==1.31.0
pandas==2.1.4
numpy==1.26.4
pyarrow==16.1.0
connectorx==0.4.3
fastexcel==0.14.0
python-calamine==0.3.2
XlsxWriter==3.2.2
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10
pyodbc==5.2.0

requests==2.32.3
beautifulsoup4==4.13.4
gspread==6.2.1

# plotly==6.2.0
# mistune==3.1.3
# schedule==1.2.2
# papermill==2.6.0
# duckdb==1.3.2
# scikit-learn==1.7.1
# ortools==9.14.6206
